oops i added the tests in the wrong place
